requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
-e git+https://github.com/pallets/flask.git@dfd3619d6f8796d48fc4e32f819cec9e8aa59156#egg=Flask
Flask-CLI==0.4.0
Flask-Testing==0.7.1
funcsigs==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==5.0.0
pathlib2==2.3.3
pluggy==0.9.0
py==1.8.0
pytest==4.4.1
requests==2.21.0
scandir==1.10.0
six==1.12.0
twill==1.8.0
urllib3==1.24.3
Werkzeug==0.15.2
